add gray2pseudo_color doc + adjust get_color_map_list function place
