+++ pdseg/__init__.py
@@ -14,3 +14,4 @@
 # limitations under the License.
 import models
 import utils
+import tools
