+++ pdseg/tools/jingling2seg.py
@@ -12,7 +12,7 @@ import numpy as np
 import PIL.Image
 import labelme
 
-from pdseg.vis import get_color_map_list
+from gray2pseudo_color import get_color_map_list
 
 
 def parse_args():
