+++ pdseg/vis.py
@@ -34,6 +34,7 @@ from utils.config import cfg
 from reader import SegDataset
 from models.model_builder import build_model
 from models.model_builder import ModelPhase
+from tools.gray2pseudo_color import get_color_map_list
 
 
 def parse_args():
@@ -71,28 +72,6 @@ def parse_args():
 def makedirs(directory):
     if not os.path.exists(directory):
         os.makedirs(directory)
-
-
-def get_color_map_list(num_classes):
-    """ Returns the color map for visualizing the segmentation mask,
-        which can support arbitrary number of classes.
-    Args:
-        num_classes: Number of classes
-    Returns:
-        The color map
-    """
-    color_map = num_classes * [0, 0, 0]
-    for i in range(0, num_classes):
-        j = 0
-        lab = i
-        while lab:
-            color_map[i * 3] |= (((lab >> 0) & 1) << (7 - j))
-            color_map[i * 3 + 1] |= (((lab >> 1) & 1) << (7 - j))
-            color_map[i * 3 + 2] |= (((lab >> 2) & 1) << (7 - j))
-            j += 1
-            lab >>= 3
-
-    return color_map
 
 
 def to_png_fn(fn):
